backend/requirements_updated.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6

# Supabase and Database
supabase==2.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Real-time and WebSocket
python-socketio==5.10.0
eventlet==0.33.3
websockets==12.0

# AI and ML services
groq==0.4.1
openai==1.3.0
transformers==4.36.0
torch==2.1.0

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.2
requests==2.31.0
validators==0.22.0
pydantic==2.5.0
python-dateutil==2.8.2
redis==5.0.1
celery==5.3.4

# Development and testing
pytest==7.4.3
pytest-flask==1.2.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and logging
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Async support
asyncio-mqtt==0.16.1
aiohttp==3.9.0

# Environment and configuration
python-decouple==3.8
environs==10.0.0

# Security
cryptography==41.0.7
PyJWT==2.8.0

# File handling
Pillow==10.1.0
python-magic==0.4.27

# Caching
cachetools==5.3.2
